requirements_frozen.txt
bidict==0.23.1
blinker==1.9.0
boto3==1.40.66
botocore==1.40.66
click==8.3.0
colorama==0.4.6
Flask==3.1.2
Flask-SocketIO==5.5.1
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
lxml==6.0.2
MarkupSafe==3.0.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-engineio==4.12.3
python-socketio==5.14.3
s3transfer==0.14.0
simple-websocket==1.1.0
six==1.17.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.2.0
